Funcionalidad para añadir reservas sobre una habitacion (#59)
* agrega propiedad monto a factura

* agrega propiedad dias de estadia con validacion de fechas

* captura el error de validacion de fechas

* comprueba si la reserva esta disponible en el rango de fechas

* upd metodo VerificarDisponibilidad para que permite que una nueva reserva comience el mismo dia en que una existente termine

+++ src/Domain/Entities/Factura/IFactura.cs
@@ -4,7 +4,7 @@
     {
         int Id { get; set; }
         DateTime Fecha { get; set; }
-        double MontoTotal { get; }
+        decimal MontoTotal { get; }
         int IdReserva { get; set; }
         Reserva Reserva { get; set; }
     }
